+++ packages/bridge-status-controller/src/bridge-status-controller.ts
@@ -74,6 +74,7 @@ import {
   handleSolanaTxResponse,
 } from './utils/transaction';
 import { generateActionId } from './utils/transaction';
+import { KeyringTypes } from '@metamask/keyring-controller';
 
 const metadata: StateMetadata<BridgeStatusControllerState> = {
   // We want to persist the bridge status state so that we can show the proper data for the Activity list
@@ -1074,10 +1075,11 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       );
     } else {
       // Submit EVM tx
-      // For hardware wallets on Mobile, this is fixes an issue where the Ledger does not get prompted for the 2nd approval
+      // For Ledger Hardware wallets on Mobile, this is fixes an issue where the Ledger does not get prompted for the 2nd approval
       // Extension does not have this issue
       const requireApproval =
-        this.#clientId === BridgeClientId.MOBILE && isHardwareAccount;
+        this.#clientId === BridgeClientId.MOBILE &&
+        selectedAccount.metadata.keyring.type === KeyringTypes.ledger;
 
       // Handle smart transactions if enabled
       txMeta = await this.#trace(