+++ packages/bridge-status-controller/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump `@metamask/keyring-api` from `^20.1.0` to `^21.0.0` ([#6560](https://github.com/MetaMask/core/pull/6560))
 
+### Fixed
+
+- Require approval for Ledger Hardware only ([#6564](https://github.com/MetaMask/core/pull/6564))
+
 ## [42.0.0]
 
 ### Added